+++ gcovr/formats/txt/write.py
@@ -54,8 +54,8 @@ def write_report(covdata: CovData, output_file: str, options: Options) -> None:
         fh.write("Directory: " + force_unix_separator(options.root) + "\n")
 
         fh.write("-" * LINE_WIDTH + "\n")
-        title_total = "Branches" if options.show_branch else "Lines"
-        title_covered = "Taken" if options.show_branch else "Exec"
+        title_total = "Branches" if options.txt_use_branch_coverage else "Lines"
+        title_covered = "Taken" if options.txt_use_branch_coverage else "Exec"
         title_percentage = "Cover"
         title_missing = "Missing"
         fh.write(
@@ -72,9 +72,10 @@ def write_report(covdata: CovData, output_file: str, options: Options) -> None:
         # Data
         keys = sort_coverage(
             covdata,
-            by_branch=options.show_branch,
+            by_branch=options.txt_use_branch_coverage,
             by_num_uncovered=options.sort_uncovered,
             by_percent_uncovered=options.sort_percent,
+            reverse=options.sort_reverse,
         )
 
         total_stat = CoverageStat.new_empty()
@@ -116,7 +117,7 @@ def write_summary_report(covdata: CovData, output_file: str, options: Options) -
 def _summarize_file_coverage(coverage: FileCoverage, options):
     filename = presentable_filename(coverage.filename, root_filter=options.root_filter)
 
-    if options.show_branch:
+    if options.txt_use_branch_coverage:
         stat = coverage.branch_coverage()
         uncovered_lines = _uncovered_branches_str(coverage)
     else: