+++ CHANGELOG.rst
@@ -26,6 +26,11 @@ New features and notable changes:
 - Add support for importing Cobertura XML files with ``--cobertura-add-tracefile`` option. (:issue:`805`)
 - Add :option:`--jacoco` to generate JaCoCo XML format. (:issue:`823`))
 - Add function coverage to HTML report. (:issue:`828`)
+- Improve sorting of data in reports (:issue:`817`):
+  - Sort file names alpha numerical and with casefold (see `str.casefold <https://docs.python.org/3.11/library/stdtypes.html?highlight=str%20casefold#str.casefold>`_) (``file_10.c`` comes after ``file_0.c``).
+  - Always sort at the end by filename if line or branch coverage is identical for a file.
+  - Add :option:`--sort-branches` to sort by branches instead of lines, this is the default if :option:`--txt-branches` is used.
+  - Add :option:`--sort-reverse` to reverse the sort order.
 
 Bug fixes and small improvements: